added 5.2 still has to test

+++ framework/shape.hpp
@@ -4,8 +4,8 @@
 class Shape
 {
 public:
-virtual int area() const = 0;
-virtual int volume() const = 0;
+virtual float area() const = 0;
+virtual float volume() const = 0;
 
 private:
